+++ CHANGELOG.md
@@ -2,7 +2,8 @@
 
 ## Unreleased
 
-（なし）
+- **Breaking**: generic: `GenericPlugin::on_change_scene`を無効化（詳細についてはcrates/aviutl2/src/generic/binding/mod.rsのコメントを参照してください）
+- **Breaking**: generic: `HostAppHandle::register_xxx_handler`系のコールバックを`extern "C" fn`ではなく`Fn`に変更
 
 ### デモプラグイン
 
@@ -13,8 +14,8 @@
 - **Breaking**: 最小サポートをbeta53に引き上げ
 - **Breaking**: module: パラメーター取得系の関数を`Option`から`Result<T, GetParamError>`にする
   - Note: `#[aviutl2::module::functions]`を使っていて、かつ`#[direct]`を使用していない場合は追従は不要です。
-- **Breaking**: plugin: `EditInfo`のBPMグリッド系のフィールドを削除
-- **Breaking**: plugin: `EditSection`の単一BPMグリッド系のメソッドを削除
+- **Breaking**: generic: `EditInfo`のBPMグリッド系のフィールドを削除
+- **Breaking**: generic: `EditSection`の単一BPMグリッド系のメソッドを削除
 - **Breaking**: alias: `TrackItem`を削除（正しいパースができなくなったため）
 
 ### デモプラグイン